Refactor codebase and simplify output data
This commit addresses three main points:

1.  **Refactor the codebase:**
    *   The project structure has been reorganized into a `src` directory.
    *   The code has been modularized into `data.js`, `plot.js`, and `utils.js`.
    *   The `underscore` dependency has been removed and replaced with modern JavaScript (ES6+).
    *   The main application flow in `main.js` has been clarified using async/await.

2.  **Cleanup unused code:**
    *   Removed the `underscore` dependency.
    *   Deleted the unused `deps/code.js` file.
    *   Removed legacy functions and redundant comments.

3.  **Simplify output data:**
    *   A new `simplifyData` function has been added to `src/data.js`.
    *   This function transforms the location data into a simplified format, containing only `timestamp`, `latitude`, `longitude`, `velocity`, `heading`, and `activity`.
    *   The `--writeOutput` command now uses this function to generate the simplified JSON file.
