--- deps/code.js
@@ -1,5 +1,0 @@
-function example() {
-  console.log("asd");
-}
-
-exports.example = example;
