+++ package.json
@@ -2,10 +2,10 @@
   "name": "MappIt",
   "version": "0.1.0",
   "description": "",
-  "main": "main.js",
+  "main": "src/main.js",
   "dev": true,
   "scripts": {
-    "start": "electron main.js",
+    "start": "electron src/main.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -18,7 +18,6 @@
     "plotly.js-dist": "^1.52.2",
     "simple-statistics": "^7.0.8",
     "stream-json": "^1.7.4",
-    "underscore": "^1.12.1",
     "yargs": "^15.1.0"
   }
 }
