+++ simplified_data.json
@@ -1,0 +1,1 @@
+{"locations":[{"timestamp":"2022-08-06T10:00:00.000Z","latitude":45.42259,"longitude":9.18641,"velocity":10,"heading":90,"activity":["IN_PASSENGER_VEHICLE"]}]}
